% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.acs.food-package.R
\docType{package}
\name{dc.acs.food-package}
\alias{dc.acs.food}
\alias{dc.acs.food-package}
\title{dc.acs.food: Social Data Commons: Health, Behavior, Diet: ACS Food Variables}
\description{
Allows user to easily get ACS food-related data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
